praying to kris jordan

+++ src/assn02/JavaWarmUp.java
@@ -48,9 +48,9 @@ public class JavaWarmUp{
 
     public static void phoneStat(Product[] database){
         int quantity = 0;
-        int totalFee = 0;
-        int totalCosts = 0;
-        int workerCosts = 0;
+        double totalFee = 0;
+        double totalCosts = 0;
+        double workerCosts = 0;
 
         for (int i = 0; i < database.length; i++) {
             if (database[i].category.equals("phone")) {
@@ -61,14 +61,14 @@ public class JavaWarmUp{
             }
         }
 
-        double avgFee = (double) totalFee/(double) quantity;
-        double avgProfits = (double) (totalFee - totalCosts - workerCosts)/(double) quantity;
+        double avgFee = totalFee/quantity;
+        double avgProfits = (totalFee - totalCosts - workerCosts)/quantity;
 
         System.out.printf("\tQuantity: %d\n\tAverage Assembling fee: %.2f\n\tAverage Net Profit: %.2f\n", quantity, avgFee, avgProfits);
     }
 
     public static void laptopStat(Product[] database){
-        double quantity = 0;
+        int quantity = 0;
         double totalFee = 0;
         double totalCosts = 0;
         double workerCosts = 0;
@@ -90,9 +90,9 @@ public class JavaWarmUp{
 
     public static void watchStat(Product[] database){
         int quantity = 0;
-        int totalFee = 0;
-        int totalCosts = 0;
-        int workerCosts = 0;
+        double totalFee = 0;
+        double totalCosts = 0;
+        double workerCosts = 0;
 
         for (int i = 0; i < database.length; i++) {
             if (database[i].category.equals("smart_watch")) {
@@ -103,8 +103,8 @@ public class JavaWarmUp{
             }
         }
 
-        double avgFee = (double) totalFee/(double) quantity;
-        double avgProfits = (double) (totalFee - totalCosts - workerCosts)/(double) quantity;
+        double avgFee = totalFee/quantity;
+        double avgProfits = (totalFee - totalCosts - workerCosts)/quantity;
 
         System.out.printf("\tQuantity: %d\n\tAverage Assembling fee: %.2f\n\tAverage Net Profit: %.2f\n", quantity, avgFee, avgProfits);
     }
